services/data-processor/pyproject.toml
[tool.poetry]
name = "data-processor"
version = "0.1.0"
description = ""
authors = ["Dudu Lasry <davidlasry696@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.5"
motor = "^3.3.2"
python-dotenv = "^1.0.1"
slack-sdk = "^3.26.2"
certifi = "^2024.2.2"
openai = "^1.12.0"
pydantic = "^2.6.1"
tqdm = "^4.66.2"
llama-index-vector-stores-pinecone = "^0.1.1"
llama-index = "^0.10.6"
llama-index-readers-slack = "^0.1.2"
llama-index-readers-notion = "^0.1.4"
llama-index-readers-github = "^0.1.4"
pygithub = "^2.2.0"
nest-asyncio = "^1.6.0"
notion-client = "^2.2.1"
google-cloud-secret-manager = "^2.18.1"
llama-index-readers-jira = "^0.1.2"
llama-index-readers-confluence = "^0.1.2"
atlassian-python-api = "^3.41.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
pytest = "^8.1.1"
pytest-sugar = "^1.0.0"
pytest-cov = "^5.0.0"
pytest-html = "^4.1.1"

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'./coverage/html' --cov-report xml:'./coverage/coverage.xml' --html='./reports/unittests/html/index.html' --junitxml='./reports/unittests/junit.xml'"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
